Add conversion and comparison operators to Timestamp.
See 135.

+++ test/t/basic/test_node.cpp
@@ -37,7 +37,7 @@ SECTION("node_builder") {
     REQUIRE(333 == node.changeset());
     REQUIRE(21 == node.uid());
     REQUIRE(std::string("foo") == node.user());
-    REQUIRE(123 == node.timestamp());
+    REQUIRE(123 == uint32_t(node.timestamp()));
     REQUIRE(osmium::Location(3.5, 4.7) == node.location());
     REQUIRE(2 == node.tags().size());
 
@@ -61,7 +61,7 @@ SECTION("node_default_attributes") {
     REQUIRE(0 == node.changeset());
     REQUIRE(0 == node.uid());
     REQUIRE(std::string("") == node.user());
-    REQUIRE(0 == node.timestamp());
+    REQUIRE(0 == uint32_t(node.timestamp()));
     REQUIRE(osmium::Location() == node.location());
     REQUIRE(0 == node.tags().size());
 }
